// functions/swagger/GripFoodBackEnd.ts
export interface Cart {
    id?: string | undefined;
    userId?: string | undefined;
    user?: User;
    restaurantId?: string | undefined;
    restaurant?: Restaurant;
    cartDetails?: CartDetail[] | undefined;
    createdAt?: Date;
}

export interface CartDetail {
    id?: string | undefined;
    cartId?: string | undefined;
    cart?: Cart;
    foodItemId?: string | undefined;
    foodItem?: FoodItem;
    qty?: number;
    createdAt?: Date;
}

export interface CheckoutModel {
    id?: string | undefined;
    name?: string | undefined;
    price?: number;
    qty?: number;
    subTotal?: number;
}

export interface FoodItem {
    id?: string | undefined;
    name?: string | undefined;
    price?: number;
    restaurantId?: string | undefined;
    restaurant?: Restaurant;
    cartDetails?: CartDetail[] | undefined;
    createdAt?: Date;
}
// ...
export interface Restaurant {
    id?: string | undefined;
    name?: string | undefined;
    foodItems?: FoodItem[] | undefined;
    carts?: Cart[] | undefined;
    createdAt?: Date;
}

export interface RestaurantCreateModel {
    name?: string | undefined;
}

export interface RestaurantUpdateModel {
    name?: string | undefined;
}

export interface User {
    id?: string | undefined;
    name?: string | undefined;
    email?: string | undefined;
    password?: string | undefined;
    carts?: Cart[] | undefined;
    createdAt?: Date;
}